chore{project-structure): moved signup files to one folder
