+++ react-app/src/App.tsx
@@ -2,7 +2,7 @@ import './App.css'
 import { BrowserRouter as Router, Routes, Route } from 'react-router-dom'
 import Home from './pages/Home'
 import Login from './pages/Login'
-import SignUp from './pages/SignUp'
+import SignUp from './pages/signup/SignUp'
 import Profile from './pages/Profile'
 import Landing from './pages/Landing'
