+++ react-app/src/pages/signup/SignUp.tsx
@@ -1,7 +1,7 @@
 // import React, { useState } from "react";
 import GoogleIcon from "@mui/icons-material/Google";
 import AppleIcon from "@mui/icons-material/Apple";
-import GatherlyLogo from "../assets/gatherlylogo.png";
+import GatherlyLogo from "../../assets/gatherlylogo.png";
 import "./SignUp.css";
 
 const SignUpPage = () => {
